# openapi.yaml
swagger: '2.0'
info:
  title: CRM Gateway Config
  description: API Gateway mapping for CRM API
  version: 1.0.0
schemes:
  - https
paths:
...
  /api/categories/{id}:
    get:
      operationId: get_id_api_categories
      parameters:
        - name: id
          in: path
          required: true
          type: string
      x-google-backend:
        address: https://crm-api-6r5doqgaaa-as.a.run.app/api/categories/{id}
      responses: { '200': { description: 'OK' } }
    put:
      operationId: put_id_api_categories
      parameters:
        - name: id
          in: path
          required: true
          type: string
      x-google-backend:
        address: https://crm-api-6r5doqgaaa-as.a.run.app/api/categories/{id}
      responses: { '200': { description: 'OK' } }
    delete:
      operationId: delete_id_api_categories
      parameters:
        - name: id
          in: path
          required: true
          type: string
      x-google-backend:
        address: https://crm-api-6r5doqgaaa-as.a.run.app/api/categories/{id}
      responses: { '200': { description: 'OK' } }
...
  /api/v1/catalog/categories/{id}:
    get:
      operationId: get_id_api_v1_catalog_categories
      parameters:
        - name: id
          in: path
          required: true
          type: string
      x-google-backend:
        address: https://crm-api-6r5doqgaaa-as.a.run.app/api/v1/catalog/categories/{id}
      responses: { '200': { description: 'OK' } }
    put:
      operationId: put_id_api_v1_catalog_categories
      parameters:
        - name: id
          in: path
          required: true
          type: string
      x-google-backend:
        address: https://crm-api-6r5doqgaaa-as.a.run.app/api/v1/catalog/categories/{id}
      responses: { '200': { description: 'OK' } }
    delete:
      operationId: delete_id_api_v1_catalog_categories
      parameters:
        - name: id
          in: path
          required: true
          type: string
      x-google-backend:
        address: https://crm-api-6r5doqgaaa-as.a.run.app/api/v1/catalog/categories/{id}
      responses: { '200': { description: 'OK' } }
...
  /api/v1/crm/categories/{id}:
    get:
      operationId: get_id_api_v1_crm_categories
      parameters:
        - name: id
          in: path
          required: true
          type: string
      x-google-backend:
        address: https://crm-api-6r5doqgaaa-as.a.run.app/api/v1/crm/categories/{id}
      responses: { '200': { description: 'OK' } }
    put:
      operationId: put_id_api_v1_crm_categories
      parameters:
        - name: id
          in: path
          required: true
          type: string
      x-google-backend:
        address: https://crm-api-6r5doqgaaa-as.a.run.app/api/v1/crm/categories/{id}
      responses: { '200': { description: 'OK' } }
    delete:
      operationId: delete_id_api_v1_crm_categories
      parameters:
        - name: id
          in: path
          required: true
          type: string
      x-google-backend:
        address: https://crm-api-6r5doqgaaa-as.a.run.app/api/v1/crm/categories/{id}
      responses: { '200': { description: 'OK' } }
...
  /api/menus/{id}:
    get:
      operationId: get_id_api_menus
      parameters:
        - name: id
          in: path
          required: true
          type: string
      x-google-backend:
        address: https://crm-api-6r5doqgaaa-as.a.run.app/api/menus/{id}
      responses: { '200': { description: 'OK' } }
    put:
      operationId: put_id_api_menus
      parameters:
        - name: id
          in: path
          required: true
          type: string
      x-google-backend:
        address: https://crm-api-6r5doqgaaa-as.a.run.app/api/menus/{id}
      responses: { '200': { description: 'OK' } }
    delete:
      operationId: delete_id_api_menus
      parameters:
        - name: id
          in: path
          required: true
          type: string
      x-google-backend:
        address: https://crm-api-6r5doqgaaa-as.a.run.app/api/menus/{id}
      responses: { '200': { description: 'OK' } }
...
  /api/v1/catalog/menus/{id}:
    get:
      operationId: get_id_api_v1_catalog_menus
      parameters:
        - name: id
          in: path
          required: true
          type: string
      x-google-backend:
        address: https://crm-api-6r5doqgaaa-as.a.run.app/api/v1/catalog/menus/{id}
      responses: { '200': { description: 'OK' } }
    put:
      operationId: put_id_api_v1_catalog_menus
      parameters:
        - name: id
          in: path
          required: true
          type: string
      x-google-backend:
        address: https://crm-api-6r5doqgaaa-as.a.run.app/api/v1/catalog/menus/{id}
      responses: { '200': { description: 'OK' } }
    delete:
      operationId: delete_id_api_v1_catalog_menus
      parameters:
        - name: id
          in: path
          required: true
          type: string
      x-google-backend:
        address: https://crm-api-6r5doqgaaa-as.a.run.app/api/v1/catalog/menus/{id}
      responses: { '200': { description: 'OK' } }
...
  /api/v1/crm/menus/{id}:
    get:
      operationId: get_id_api_v1_crm_menus
      parameters:
        - name: id
          in: path
          required: true
          type: string
      x-google-backend:
        address: https://crm-api-6r5doqgaaa-as.a.run.app/api/v1/crm/menus/{id}
      responses: { '200': { description: 'OK' } }
    put:
      operationId: put_id_api_v1_crm_menus
      parameters:
        - name: id
          in: path
          required: true
          type: string
      x-google-backend:
        address: https://crm-api-6r5doqgaaa-as.a.run.app/api/v1/crm/menus/{id}
      responses: { '200': { description: 'OK' } }
    delete:
      operationId: delete_id_api_v1_crm_menus
      parameters:
        - name: id
          in: path
          required: true
          type: string
      x-google-backend:
        address: https://crm-api-6r5doqgaaa-as.a.run.app/api/v1/crm/menus/{id}
      responses: { '200': { description: 'OK' } }
...
  /api/brands/{id}:
    get:
      operationId: get_id_api_brands
      parameters:
        - name: id
          in: path
          required: true
          type: string
      x-google-backend:
        address: https://crm-api-6r5doqgaaa-as.a.run.app/api/brands/{id}
      responses: { '200': { description: 'OK' } }
    put:
      operationId: put_id_api_brands
      parameters:
        - name: id
          in: path
          required: true
          type: string
      x-google-backend:
        address: https://crm-api-6r5doqgaaa-as.a.run.app/api/brands/{id}
      responses: { '200': { description: 'OK' } }
    delete:
      operationId: delete_id_api_brands
      parameters:
        - name: id
          in: path
          required: true
          type: string
      x-google-backend:
        address: https://crm-api-6r5doqgaaa-as.a.run.app/api/brands/{id}
      responses: { '200': { description: 'OK' } }
...
  /api/v1/catalog/brands/{id}:
    get:
      operationId: get_id_api_v1_catalog_brands
      parameters:
        - name: id
          in: path
          required: true
          type: string
      x-google-backend:
        address: https://crm-api-6r5doqgaaa-as.a.run.app/api/v1/catalog/brands/{id}
      responses: { '200': { description: 'OK' } }
    put:
      operationId: put_id_api_v1_catalog_brands
      parameters:
        - name: id
          in: path
          required: true
          type: string
      x-google-backend:
        address: https://crm-api-6r5doqgaaa-as.a.run.app/api/v1/catalog/brands/{id}
      responses: { '200': { description: 'OK' } }
    delete:
      operationId: delete_id_api_v1_catalog_brands
      parameters:
        - name: id
          in: path
          required: true
          type: string
      x-google-backend:
        address: https://crm-api-6r5doqgaaa-as.a.run.app/api/v1/catalog/brands/{id}
      responses: { '200': { description: 'OK' } }
...
  /api/v1/crm/brands/{id}:
    get:
      operationId: get_id_api_v1_crm_brands
      parameters:
        - name: id
          in: path
          required: true
          type: string
      x-google-backend:
        address: https://crm-api-6r5doqgaaa-as.a.run.app/api/v1/crm/brands/{id}
      responses: { '200': { description: 'OK' } }
    put:
      operationId: put_id_api_v1_crm_brands
      parameters:
        - name: id
          in: path
          required: true
          type: string
      x-google-backend:
        address: https://crm-api-6r5doqgaaa-as.a.run.app/api/v1/crm/brands/{id}
      responses: { '200': { description: 'OK' } }
    delete:
      operationId: delete_id_api_v1_crm_brands
      parameters:
        - name: id
          in: path
          required: true
          type: string
      x-google-backend:
        address: https://crm-api-6r5doqgaaa-as.a.run.app/api/v1/crm/brands/{id}
      responses: { '200': { description: 'OK' } }
...
  /api/modifiergroups/{id}:
    get:
      operationId: get_id_api_modifiergroups
      parameters:
        - name: id
          in: path
          required: true
          type: string
      x-google-backend:
        address: https://crm-api-6r5doqgaaa-as.a.run.app/api/modifiergroups/{id}
      responses: { '200': { description: 'OK' } }
    put:
      operationId: put_id_api_modifiergroups
      parameters:
        - name: id
          in: path
          required: true
          type: string
      x-google-backend:
        address: https://crm-api-6r5doqgaaa-as.a.run.app/api/modifiergroups/{id}
      responses: { '200': { description: 'OK' } }
    delete:
      operationId: delete_id_api_modifiergroups
      parameters:
        - name: id
          in: path
          required: true
          type: string
      x-google-backend:
        address: https://crm-api-6r5doqgaaa-as.a.run.app/api/modifiergroups/{id}
      responses: { '200': { description: 'OK' } }
...
  /api/v1/catalog/modifiergroups/{id}:
    get:
      operationId: get_id_api_v1_catalog_modifiergroups
      parameters:
        - name: id
          in: path
          required: true
          type: string
      x-google-backend:
        address: https://crm-api-6r5doqgaaa-as.a.run.app/api/v1/catalog/modifiergroups/{id}
      responses: { '200': { description: 'OK' } }
    put:
      operationId: put_id_api_v1_catalog_modifiergroups
      parameters:
        - name: id
          in: path
          required: true
          type: string
      x-google-backend:
        address: https://crm-api-6r5doqgaaa-as.a.run.app/api/v1/catalog/modifiergroups/{id}
      responses: { '200': { description: 'OK' } }
    delete:
      operationId: delete_id_api_v1_catalog_modifiergroups
      parameters:
        - name: id
          in: path
          required: true
          type: string
      x-google-backend:
        address: https://crm-api-6r5doqgaaa-as.a.run.app/api/v1/catalog/modifiergroups/{id}
      responses: { '200': { description: 'OK' } }
...
  /api/v1/crm/modifiergroups/{id}:
    get:
      operationId: get_id_api_v1_crm_modifiergroups
      parameters:
        - name: id
          in: path
          required: true
          type: string
      x-google-backend:
        address: https://crm-api-6r5doqgaaa-as.a.run.app/api/v1/crm/modifiergroups/{id}
      responses: { '200': { description: 'OK' } }
    put:
      operationId: put_id_api_v1_crm_modifiergroups
      parameters:
        - name: id
          in: path
          required: true
          type: string
      x-google-backend:
        address: https://crm-api-6r5doqgaaa-as.a.run.app/api/v1/crm/modifiergroups/{id}
      responses: { '200': { description: 'OK' } }
    delete:
      operationId: delete_id_api_v1_crm_modifiergroups
      parameters:
        - name: id
          in: path
          required: true
          type: string
      x-google-backend:
        address: https://crm-api-6r5doqgaaa-as.a.run.app/api/v1/crm/modifiergroups/{id}
      responses: { '200': { description: 'OK' } }
...